Move lightclient code to package root
